+++ models/ordered_item_view.go
@@ -11,9 +11,9 @@ type OrderedItemView struct {
 	ProductID        string                 `json:"product_id"`
 	Name             string                 `json:"name"`
 	Quantity         int                    `json:"quantity"`
-	Price            int                    `json:"price"`
-	ProductCost      int                    `json:"product_cost"`
-	SubTotal         int                    `json:"sub_total"`
+	Price            int64                  `json:"price"`
+	ProductCost      int64                  `json:"product_cost"`
+	SubTotal         int64                  `json:"sub_total"`
 	Description      string                 `json:"description"`
 	SKU              string                 `json:"sku"`
 	AdditionalImages string                 `json:"additional_images"`
