+++ models/shipping_method.go
@@ -6,7 +6,7 @@ type ShippingMethod struct {
 	ID                      string    `json:"id" sql:"id" gorm:"primary_key"`
 	Name                    string    `json:"name" sql:"name" gorm:"unique;not null"`
 	ApproximateDeliveryTime int       `json:"approximate_delivery_time" gorm:"approximate_delivery_time" gorm:"index"`
-	DeliveryCharge          int       `json:"delivery_charge" sql:"delivery_charge" gorm:"index"`
+	DeliveryCharge          int64     `json:"delivery_charge" sql:"delivery_charge" gorm:"index"`
 	WeightUnit              string    `json:"weight_unit" sql:"weight_unit"`
 	IsFlat                  bool      `json:"is_flat" gorm:"column:is_flat"`
 	IsPublished             bool      `json:"is_published" sql:"is_published" gorm:"index"`
@@ -18,7 +18,7 @@ func (sm *ShippingMethod) TableName() string {
 	return "shipping_methods"
 }
 
-func (sm *ShippingMethod) CalculateDeliveryCharge(weight int) int {
+func (sm *ShippingMethod) CalculateDeliveryCharge(weight int) int64 {
 	// TODO : Currently weight doesn't have any impact on charge
 	return sm.DeliveryCharge
 }
